+++ src/main/java/moze_intel/projecte/emc/mappers/CustomEMCMapper.java
@@ -8,10 +8,10 @@ import net.minecraftforge.common.config.Configuration;
 
 import java.util.Map;
 
-public class CustomEMCMapper implements IEMCMapper<NormalizedSimpleStack, Integer> {
+public class CustomEMCMapper implements IEMCMapper<NormalizedSimpleStack, Long> {
 	@Override
-	public void addMappings(IMappingCollector<NormalizedSimpleStack, Integer> mapper, Configuration config) {
-		for (Map.Entry<NormalizedSimpleStack,Integer> entry : CustomEMCParser.userValues.entrySet()) {
+	public void addMappings(IMappingCollector<NormalizedSimpleStack, Long> mapper, Configuration config) {
+		for (Map.Entry<NormalizedSimpleStack,Long> entry : CustomEMCParser.userValues.entrySet()) {
 			PELogger.logInfo("Adding custom EMC value for " + entry.getKey() + ": " + entry.getValue());
 			mapper.setValueBefore(entry.getKey(), entry.getValue());
 		}
