+++ src/main/java/moze_intel/projecte/gameObjs/gui/GUICollectorMK2.java
@@ -57,7 +57,8 @@ public class GUICollectorMK2 extends GuiContainer
 		this.drawTexturedModalRect(x + 80, y + 58, 0, 166, progress, 10);
 		
 		//Fuel Progress. Max is 24.
-		progress = tile.getFuelProgressScaled(24);
+		//TODO: long
+		progress = (int) tile.getFuelProgressScaled(24);
 		this.drawTexturedModalRect(x + 154, y + 55 - progress, 201, 38 - progress, 10, progress + 1);
 	}
 }
