+++ src/main/java/moze_intel/projecte/gameObjs/gui/GUICollectorMK3.java
@@ -58,7 +58,8 @@ public class GUICollectorMK3 extends GuiContainer
 		this.drawTexturedModalRect(x + 98, y + 58, 0, 166, progress, 10);
 		
 		//Fuel Progress. Max is 24.
-		progress = tile.getFuelProgressScaled(24);
+		//TODO: long
+		progress = (int) tile.getFuelProgressScaled(24);
 		this.drawTexturedModalRect(x + 172, y + 55 - progress, 219, 38 - progress, 10, progress + 1);
 	}
 }
